Added a checked next_states implementation
The initial state of a Combinations type is a tuple of the first K
states of the wrapped iterator. However, there is no guarantee that this
iterator will be at least K elements long.
The start(::Combinations) method now uses this implementation. This will
make sure that next is never called with a 'done state'.